+++ semcore/utils/src/inputProps.ts
@@ -19,6 +19,7 @@ export const inputProps = [
   'onKeyUp',
   'tabIndex',
   'data-ui-name',
+  'inputMode',
 ];
 
 type ExtractedProps<T extends object, K extends Array<keyof T>> = Partial<