+++ semcore/time-picker/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.49.10] - 2025-01-21
+
+### Added
+
+- `inputmode=numeric` for inputs.
+
+## [4.49.9] - 2025-01-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.40.2 ~> 5.40.3], `@semcore/select` [4.49.1 ~> 4.49.2]).
+
 ## [4.49.8] - 2024-12-30
 
 ### Changed